% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmischelpers.R
\name{add_leading_zeros}
\alias{add_leading_zeros}
\title{add_leading_zeros}
\usage{
add_leading_zeros(geoid_type = "zip", input)
}
\arguments{
\item{geoid_type}{What geoid type to add leading 0s to
\enumerate{
\item zip -> Must be 5 digit zip code
\item county -> Must be a 5 digit county fips code
\item cbsa -> Must be a 5 digit cbsa code
\item cbsadiv -> Must be a 5 digit cbsadiv code
\item cd -> Must be a 4 digit cd code.
If you specify a number, it will
make the geoid of that length.
}}

\item{input}{A vector or column of character vectors that need
leading zeros to be processed properly.}
}
\value{
A vector with leading zeros added.
}
\description{
Sometimes when loading in data from an excel file it
truncates the leading 0's. This attempts to re-add those.
}
\examples{
\dontrun{
library(hudr)
zip <- c(02102, 11032, 01232)

zip <- hud_add_leading_zeros("zip", zip)

cbsa <- c(02102, 11032, 01232)

cbsa <- hud_add_leading_zeros("cbsa", cbsa)

some_geoid <- c(213113, 6526266, 1316455)

some_geoid <- hud_add_leading_zeros(10, some_geoid)
}
}
