% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudparallel.R
\name{api_call}
\alias{api_call}
\title{api_call}
\usage{
api_call(listed)
}
\arguments{
\item{listed}{Each individual entry url to query for as well as index an meta data.}
}
\value{
The dataframe collection of all the response bodies.
}
\description{
Make API calls
}
\keyword{call}
