% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_in_geo.R
\name{cbsadiv_in_z}
\alias{cbsadiv_in_z}
\title{cbsadiv_in_z}
\usage{
cbsadiv_in_z(cbsadiv, zip, year, quarter)
}
\arguments{
\item{cbsadiv}{The cbsadiv to determine overlap with zip}

\item{zip}{The zip to determine overlap with cbsadiv}

\item{year}{The year of the crosswalk files.}

\item{quarter}{The quarter of the crosswalk files.}
}
\description{
Given a cbsadiv and a zip, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

z_in_cbsadiv(cbsadiv = 43340, zip = 71052, year = year, quarter = quarter)

}
}
\seealso{
\itemize{
\item \code{\link[=z_in_trt]{z_in_trt()}}
\item \code{\link[=z_in_county]{z_in_county()}}
\item \code{\link[=z_in_cbsa]{z_in_cbsa()}}
\item \code{\link[=z_in_cbsadiv]{z_in_cbsadiv()}}
\item \code{\link[=z_in_ctysb]{z_in_ctysb()}}
\item \code{\link[=z_cd]{z_cd()}}
\item \code{\link[=trt_in_z]{trt_in_z()}}
\item \code{\link[=cty_in_z]{cty_in_z()}}
\item \code{\link[=cbsa_in_z]{cbsa_in_z()}}
\item \code{\link[=cbsadiv_in_z]{cbsadiv_in_z()}}
\item \code{\link[=cd_in_z]{cd_in_z()}}
\item \code{\link[=ctysb_in_z]{ctysb_in_z()}}
}
}
