% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudcrosswalkadataset.R
\name{crosswalk}
\alias{crosswalk}
\title{crosswalk}
\usage{
crosswalk(
  dataset,
  type,
  geoid_col,
  crosswalkable_fields = NA,
  ratio = "tot",
  round = FALSE,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{dataset}{The dataset}

\item{type}{The crosswalk types:
\enumerate{
\item zip-tract
\item zip-county
\item zip-cbsa
\item zip-cbsadiv (Available 4th Quarter 2017 onwards)
\item zip-cd
\item tract-zip
\item county-zip
\item cbsa-zip
\item cbsadiv-zip (Available 4th Quarter 2017 onwards)
\item cd-zip
\item zip-countysub (Available 2nd Quarter 2018 onwards)
\item countysub-zip (Available 2nd Quarter 2018 onwards)
}}

\item{geoid_col}{The column where the geoid is located.}

\item{crosswalkable_fields}{The fields in dataset that are able to be
crosswalked. Should only be numeric values in them. If not specified, will
assume all fields that are of type numeric double or integer are
crosswalkable. This may be problematic if numbers are used as categorical
data such as a customer ID. You can choose to not include or include columns.}

\item{ratio}{Can either choose residential, business, others, or total
ratios. If one is supplied it will apply to all crosswalk able fields. If
the same length is provided, will crosswalk it with those specific ratios
for each column respectively. Defaults to the total ratio of buildings.
\enumerate{
\item "res"
\item "bus"
\item "oth"
\item "tot"
}}

\item{round}{Since numbers are multiplied by a ratio, discrete integers will
become continuous.}

\item{year}{The years to query for.}

\item{quarter}{The quarters in the year to query for.}

\item{key}{The key obtained at HUD User.}
}
\value{
The entire dataset crosswalked to a different geoid.
}
\description{
Will crosswalk a dataset given the geoid column and fields that
should be crosswalked. This function does not handle large crosswalks very
well. Performance enhancements have not been implemented.
}
