% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudparallel.R
\name{determine_optimal_core_count}
\alias{determine_optimal_core_count}
\title{determine_optimal_core_count}
\usage{
determine_optimal_core_count()
}
\value{
The number of optimal cores for this user
}
\description{
Determines the optimal number of cores to use when doing parallel API calls.
}
\keyword{cores}
