% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfixfips.R
\name{fix_county_fips}
\alias{fix_county_fips}
\title{fix_county_fips}
\usage{
fix_county_fips(county_fips)
}
\arguments{
\item{county_fips}{A column or vector of fips data which should be 3 characters long.}
}
\value{
A dataframe with the corrected county fips.
}
\description{
Sometimes when loading in fips data from an excel file it truncates the leading 0's. This attempts to re-add those.
}
