% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfixfips.R
\name{fix_fips}
\alias{fix_fips}
\title{fix_fips}
\usage{
fix_fips(fips)
}
\arguments{
\item{fips}{A column or vector of fips data which should be 5 characters long.}
}
\value{
A dataframe with the corrected fips.
}
\description{
Sometimes when loading in fips data from an excel file it truncates the leading 0's. This attempts to re-add those.
}
