% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfixfips.R
\name{fix_state_fips}
\alias{fix_state_fips}
\title{fix_state_fips}
\usage{
fix_state_fips(state_fips)
}
\arguments{
\item{state_fips}{A column or vector of fips data which should be 2 characters long.}
}
\value{
A dataframe with only the corrected state fips.
}
\description{
Sometimes when loading in fips data from an excel file it truncates the leading 0's. This attempts to re-add those.
}
