% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{geo_is_infix_query_and_get_warnings}
\alias{geo_is_infix_query_and_get_warnings}
\title{geo_is_infix_query_and_get_warnings}
\usage{
geo_is_infix_query_and_get_warnings(query, f, year, quarter, querytype)
}
\arguments{
\item{query}{The geoids to query for crosswalk}

\item{f}{The function used query the crosswalk files.}

\item{year}{The year to query for.}

\item{quarter}{The quarter to query for.}
}
\description{
geo_is_infix_query_and_get_warnings
}
