% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{geo_is_infix_rhs_cleansing}
\alias{geo_is_infix_rhs_cleansing}
\title{geo_is_infix_rhs_cleansing}
\usage{
geo_is_infix_rhs_cleansing(query, geoid_type)
}
\arguments{
\item{query}{The geoid to query for}

\item{geoid_type}{The type of geoid, either:
\enumerate{
\item zip,
\item tract,
\item cd,
\item cbsa,
\item cbsadiv,
\item county,
\item countysub
}}
}
\description{
Given a geographic identifier, described by the
crosswalk files, determine whether it is the right length and has
correct spacing.
}
