% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_host_name.R
\name{get_hud_host_name}
\alias{get_hud_host_name}
\title{Base URL for HUD USER}
\usage{
get_hud_host_name()
}
\description{
Returns the hardcoded base URL location for HUD USER.
}
\examples{
\dontrun{

get_hud_host_name()

}
}
