% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%cbsa_in_z\%}
\alias{\%cbsa_in_z\%}
\title{\%cbsa_in_z\%}
\usage{
cbsa \%cbsa_in_z\% zip
}
\arguments{
\item{cbsa}{The cbsa to determine overlap with zip}

\item{zip}{The zip to determine overlap with cbsa}
}
\description{
Given a cbsa and a zip, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

43340 \%cbsa_in_z\% 71052

}
}
