% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%cd_in_z\%}
\alias{\%cd_in_z\%}
\title{\%cd_in_z\%}
\usage{
cd \%cd_in_z\% zip
}
\arguments{
\item{cd}{The cd to determine overlap with zip}

\item{zip}{The zip to determine overlap with cd}
}
\description{
Given a congressional district and a zip, determine if they
overlap using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.
}
\examples{
\dontrun{

2204 \%cd_in_z\% 71052

}
}
