% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%ctysb_in_z\%}
\alias{\%ctysb_in_z\%}
\title{\%ctysb_in_z\%}
\usage{
countysub \%ctysb_in_z\% zip
}
\arguments{
\item{zip}{The zip to determine overlap with countysub}

\item{ctysb}{The countysub to determine overlap with zip}
}
\description{
Given a countysub and zip code, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

2203194756 \%ctysb_in_z\% 71052

}
}
