% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_cbsadiv\%}
\alias{\%z_in_cbsadiv\%}
\title{\%z_in_cbsadiv\%}
\usage{
zip \%z_in_cbsadiv\% cbsadiv
}
\arguments{
\item{zip}{The zip to determine overlap with cbsadiv}

\item{cbsadiv}{The cbsadiv to determine overlap with zip}
}
\description{
Given a zip code and a cbsadiv, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

71052 \%z_in_cbsadiv\% 43340

}
}
