% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_cd\%}
\alias{\%z_in_cd\%}
\title{\%z_in_cd\%}
\usage{
zip \%z_in_cd\% cd
}
\arguments{
\item{zip}{The zip to determine overlap with a congressional district}

\item{cd}{The congressional district to determine overlap with zip}
}
\description{
Given a zip code and a congressional district, determine if
they overlap using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

71052 \%z_in_cd\% 2204

}
}
