% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_cty\%}
\alias{\%z_in_cty\%}
\title{\%z_in_cty\%}
\usage{
zip \%z_in_cty\% county
}
\arguments{
\item{zip}{The zip to determine overlap with tract}

\item{county}{The county to determine overlap with zip}
}
\description{
Given a zip code and a county, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.
}
\examples{
\dontrun{

71052 \%z_in_cty\% 22031

}
}
