% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_ctysb\%}
\alias{\%z_in_ctysb\%}
\title{\%z_in_ctysb\%}
\usage{
zip \%z_in_ctysb\% countysub
}
\arguments{
\item{zip}{The zip to determine overlap with countysub}

\item{countysub}{The countysub to determine overlap with zip}
}
\description{
Given a zip code and a countysub, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

71052 \%z_in_ctysb\% 2203194756

}
}
