% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_trt\%}
\alias{\%z_in_trt\%}
\title{\%z_in_trt\%}
\usage{
zip \%z_in_trt\% tract
}
\arguments{
\item{zip}{The zip to determine overlap with tract}

\item{tract}{The tract to determine overlap with zip}
}
\description{
Given a zip code and a tract, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

71052 \%z_in_trt\% 22031950600

}
}
