% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_50p}
\alias{hud_50p}
\title{hud_50p}
\usage{
hud_50p(year = 2022, res = "c")
}
\arguments{
\item{year}{The specific year for the 50th percentile (or median) to query for.}

\item{res}{Can choose between county or FMR area resolution for 50\% percentile.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the 50th percentile (or median)
}
