% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_adjf}
\alias{hud_adjf}
\title{hud_adjf}
\usage{
hud_adjf(year = 2021, sheet = 1)
}
\arguments{
\item{year}{The specific year for the Annual Adjustment Factors to query for.}

\item{sheet}{The specific sheet in the file.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the
Annual Adjustment Factors dataset.
}
