% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_ash}
\alias{hud_ash}
\title{hud_ash}
\usage{
hud_ash(year = 2021, res = 1)
}
\arguments{
\item{year}{The specific year for the Picture of Subsidized Households.
Supports years 2021 to 2009.}

\item{res}{The GEOID or place resolution(2012-2021 are based
2010 census geographies)
\enumerate{
\item U.S. Total
\item State
\item Core Based Statistical Areas
\item Public Housing Agency
\item Project
\item Census Tract: AK - MN
\item Census Tract: MO - WY
\item City
\item County
\item Congressional District
\item Zipcode
}}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Picture of
Subsidized Households.
}
