% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudexchange.R
\name{hud_cdbg}
\alias{hud_cdbg}
\title{hud_cdbg}
\usage{
hud_cdbg(file = 1)
}
\arguments{
\item{file}{The specific file needed. Currently you can specify 1 -> 4. Will default to 1:
\enumerate{
\item DR CDBG CDBG-DR Financial Report by Appropriation (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Financial Report by Grantee (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Financial Report Monthly Summary (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Performance by Activity (xlsx - 2.2 MB - 1/1/2022)
}}
}
\value{
A dataframe.
}
\description{
This will grab data from
https://drgr.hud.gov/public/data_downloads.html?programName=DR\%20CDBG
and format it into a dataframe. It will unroll each column to replace
NA values with the first non-NA value above it.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
hud_cdbg(1)

hud_cdbg(2)

hud_cdbg(file = 3)

hud_cdbg(file = 4)
}
}
