% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudchas.R
\name{hud_chas_county}
\alias{hud_chas_county}
\title{hud_chas_county}
\usage{
hud_chas_county(
  county,
  year = format(Sys.Date(), "\%Y"),
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{county}{The county to query for. Must supply a geoid. 2 digit state fip + 3 digit county fip. hud_counties() will show an extra 99999 at the end. Just remove that.}

\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtain from HUD USER website.}
}
\value{
Returns a dataframe with CHAS data for a particular county.
}
\description{
Returns CHAS data for a county.
}
