% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudchas.R
\name{hud_chas_nation}
\alias{hud_chas_nation}
\title{hud_chas_nation}
\usage{
hud_chas_nation(year = format(Sys.Date(), "\%Y"), key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtain from HUD USER website.}
}
\value{
Returns a dataframe with CHAS data for the entire nation.
}
\description{
Returns CHAS data for the entire nation.
}
