% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_chas.R
\name{hud_chas_state}
\alias{hud_chas_state}
\title{hud_chas_state}
\usage{
hud_chas_state(
  state,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{state}{The state to query for. Can supply as abbreviation, whole name,
or as geoid.}

\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtain from HUD USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
Returns a data frame with CHAS data for states for all combinations
of "state" and "year" inputs.
}
\description{
Returns CHAS data for a state.
}
\examples{
\dontrun{
hud_chas_state("CA")
hud_chas_state("51")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_chas_nation]{hud_chas_nation()}}
\item \code{\link[=hud_chas_state]{hud_chas_state()}}
\item \code{\link[=hud_chas_county]{hud_chas_county()}}
\item \code{\link[=hud_chas_state_mcd]{hud_chas_state_mcd()}}
\item \code{\link[=hud_chas_state_place]{hud_chas_state_place()}}
}
}
