% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_chas.R
\name{hud_chas_state_place}
\alias{hud_chas_state_place}
\title{Comprehensive Housing and Affordability Strategy for US Places}
\usage{
hud_chas_state_place(
  state,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble = getOption("rhud_use_tibble", FALSE)
)
}
\arguments{
\item{state}{A character or numeric vector:
The states to query for which can be specified using
the state name, abbreviation, or fips code.}

\item{year}{A character vector: the years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{A character vector of length one with the key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{A logical : if TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
Returns a dataframe with
Comprehensive Housing and Affordability Strategy (CHAS) data for places
inside states for all
combinations of "state" and "year" inputs.
}
\description{
Returns Comprehensive Housing and Affordability Strategy (CHAS)
for all places in a state.
}
\examples{
\dontrun{

#hud_chas_state_place("MD", year = c("2014-2018","2013-2017"))

}
}
\seealso{
\itemize{
\item \code{\link[=hud_chas_nation]{hud_chas_nation()}}
\item \code{\link[=hud_chas_state]{hud_chas_state()}}
\item \code{\link[=hud_chas_county]{hud_chas_county()}}
\item \code{\link[=hud_chas_state_mcd]{hud_chas_state_mcd()}}
\item \code{\link[=hud_chas_state_place]{hud_chas_state_place()}}
\item \code{\link[=hud_chas]{hud_chas()}}
}
}
