% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_cities}
\alias{hud_cities}
\title{hud_cities}
\usage{
hud_cities(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all cities}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing details of cities in a state
}
\description{
Get a list of all cities in a state
}
\keyword{Cities}
