% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_cache.R
\name{hud_clear_cache}
\alias{hud_clear_cache}
\title{hud_clear_cache}
\usage{
hud_clear_cache()
}
\description{
Remove cached data from the caching directory that is used to
store data retrieved using the rhud API calls. By default, rhud uses a
non-persistent temporary directory given for an R session, but a user might
have set another directory to use.
}
\examples{
\dontrun{

hud_clear_cache()

}
}
\seealso{
\itemize{
\item \code{\link[=hud_get_cache_dir]{hud_get_cache_dir()}}
\item \code{\link[=hud_set_cache_dir]{hud_set_cache_dir()}}
\item \code{\link[=hud_clear_cache]{hud_clear_cache()}}
}
}
