% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_counties}
\alias{hud_counties}
\title{hud_counties}
\usage{
hud_counties(state)
}
\value{
A dataframe containing all counties within a state
}
\description{
Get a list of all counties within a state
}
\keyword{Counties}
