% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huduser.R
\name{hud_cw}
\alias{hud_cw}
\title{hud_cw}
\usage{
hud_cw(
  type,
  query,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{type}{Must be a number between 1 and 12 depending on the Crosswalk
type. You can also supply the string name.
\enumerate{
\item zip-tract
\item zip-county
\item zip-cbsa
\item zip-cbsadiv (Available 4th Quarter 2017 onwards)
\item zip-cd
\item tract-zip
\item county-zip
\item cbsa-zip
\item cbsadiv-zip (Available 4th Quarter 2017 onwards)
\item cd-zip
\item zip-countysub (Available 2nd Quarter 2018 onwards)
\item countysub-zip (Available 2nd Quarter 2018 onwards)
}}

\item{query}{5 digit USPS ZIP code of the data to retrieve.
E.g. 22031 for type 1 to 5 and 11 .
or
11 digit unique 2000 or 2010 Census tract GEOID consisting of
state FIPS + county FIPS + tract code. Eg: 51059461700  for type 6
or
5 digit unique 2000 or 2010 Census county GEOID consisting of
state FIPS + county FIPS. Eg: 51600 for type 7
or
5 digit CBSA code for Micropolitan and Metropolitan Areas
Eg: 10380 for type 8
or
5-digit CBSA Division code which only applies to Metropolitan Areas.
Eg: 35614 for type 9
or
4-digit GEOID for the Congressional District which consists of
state FIPS + Congressional District code. Eg: 7200 for type 10
or
10-digit GEOID for the County sub Eg: 4606720300 for type 12}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{quarter}{Gets the quarter of the year that this data was recorded.
Defaults to the first quarter of the year.}

\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return
all fields. This does not remove duplicates.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
This function returns a dataframe containing CROSSWALK data for
a particular GEOID. These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the Crosswalks API provided by
US Department of Housing and Urban Development.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_cw(type = 1, query = '35213', year = c('2010', '2011'),
   quarter = c('1'))

hud_cw(type = '2', query = '35213', year = c('2016', '2020'),
   quarter = c('2'))

hud_cw(type = 3, query = 35213, year = c('2012', '2011'),
   quarter = c('3'))

hud_cw(type = 4, query = '22031', year = c('2017', '2019'),
   quarter = c('4'))

hud_cw(type = '5', query = '35213', year = c(2011, '2012'),
   quarter = c('1','2'))

hud_cw(type = 6, query = '48201223100', year = c('2017', '2010'),
   quarter = c('1','2','3'))

hud_cw(type = 7, query = '22031', year = c('2010', '2011'),
   quarter = c('1','2','3','4'))

hud_cw(type = 8, query = '10140', year = c('2010', '2011'),
   quarter = c('2','1'))

hud_cw(type = 9, query = '10380', year = c('2017'),
   quarter = c('4'))

hud_cw(type = 10, query = '2202', year = c('2010', '2011'),
   quarter = c('4','3'))

hud_cw(type = 11, query = '35213', year = c('2019', '2020'),
   quarter = c('2','3'))

hud_cw(type = 12, query = '4606720300 ', year = c('2019', '2019', '2019'),
   quarter = c('4','4'))
}
}
\keyword{API}
\keyword{Crosswalks}
