% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudcw.R
\name{hud_cw_county_zip}
\alias{hud_cw_county_zip}
\title{hud_cw_county_zip}
\usage{
hud_cw_county_zip(
  county,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{county}{5 digit unique 2000 or 2010 Census county GEOID consisting of
state FIPS + county FIPS. Eg: 51600 for type 7}

\item{year}{Gets the year that this data was recorded. Can specify multiple
years. Default is the previous year.}

\item{quarter}{Gets the quarter of the year that this data was recorded.
Defaults to the first quarter of the year.}

\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return
all fields.}

\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}
}
\value{
This function returns a dataframe containing CROSSWALK data for
a particular GEOID. These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the Crosswalks API provided by
US Department of Housing and Urban Development. This
returns the crosswalk for county to zip.
}
\examples{
\dontrun{
library(hudr)

hud_cw_county_zip(county = 22031, year = c('2017'), quarter = c('1'))

hud_cw_county_zip(county = '22031', year = c('2010'), quarter = c('1'),
   minimal = TRUE)
}
}
\keyword{API}
\keyword{Crosswalks}
