% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudcw.R
\name{hud_cw_countysub_zip}
\alias{hud_cw_countysub_zip}
\title{hud_cw_countysub_zip}
\usage{
hud_cw_countysub_zip(
  countysub,
  year = format(Sys.Date(), "\%Y"),
  quarter = 1,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{countysub}{10-digit GEOID for the County sub Eg: 4606720300 for type 12}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
current year.}

\item{quarter}{Gets the quarter of the year that this data was recorded.
Defaults to the first quarter of the year.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
This function returns a dataframe containing CROSSWALK data for
a particular GEOID. These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the Crosswalks API provided by
US Department of Housing and Urban Development. This
returns the crosswalk for countysub to zip.
}
\keyword{API}
\keyword{Crosswalks}
