% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_crosswalk.R
\name{hud_cw_tract_zip}
\alias{hud_cw_tract_zip}
\title{hud_cw_tract_zip}
\usage{
hud_cw_tract_zip(
  tract,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{tract}{11 digit unique 2000 or 2010 Census tract GEOID consisting of
state FIPS + county FIPS + tract code. Eg: 51059461700  for type 6}

\item{year}{Gets the year that this data was recorded. Can specify multiple
years. Default is the previous year.}

\item{quarter}{Gets the quarter of the year that this data was recorded.
Defaults to the first quarter of the year.}

\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return
all fields. This does not remove duplicates.}

\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
This function returns a dataframe containing crosswalk data for
tract to zip for all combinations of "tract", "year", and "quarter"
inputs.

These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the Crosswalks API provided by US
Department of Housing and Urban Development. This returns the crosswalk for
tract to zip.
}
\examples{
\dontrun{
library(rhud)

hud_cw_tract_zip(tract = 48201223100, year = c('2017'), quarter = c('1'))

hud_cw_tract_zip(tract = '48201223100', year = c('2010'), quarter = c('1'),
   minimal = TRUE)
}
}
\seealso{
\itemize{
\item \code{\link[=hud_cw_zip_tract]{hud_cw_zip_tract()}}
\item \code{\link[=hud_cw_zip_county]{hud_cw_zip_county()}}
\item \code{\link[=hud_cw_zip_cbsa]{hud_cw_zip_cbsa()}}
\item \code{\link[=hud_cw_zip_cbsadiv]{hud_cw_zip_cbsadiv()}}
\item \code{\link[=hud_cw_zip_countysub]{hud_cw_zip_countysub()}}
\item \code{\link[=hud_cw_zip_cd]{hud_cw_zip_cd()}}
\item \code{\link[=hud_cw_tract_zip]{hud_cw_tract_zip()}}
\item \code{\link[=hud_cw_county_zip]{hud_cw_county_zip()}}
\item \code{\link[=hud_cw_cbsa_zip]{hud_cw_cbsa_zip()}}
\item \code{\link[=hud_cw_cbsadiv_zip]{hud_cw_cbsadiv_zip()}}
\item \code{\link[=hud_cw_cd_zip]{hud_cw_cd_zip()}}
\item \code{\link[=hud_cw_countysub_zip]{hud_cw_countysub_zip()}}
\item \code{\link[=hud_cw]{hud_cw()}}
}
}
\keyword{API}
\keyword{Crosswalks}
