% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{hud_data_sets}
\alias{hud_data_sets}
\title{hud_data_sets}
\usage{
hud_data_sets()
}
\value{
Dataframe with information about the various datasets that HUD has,
update information, and corresponding functions/documentation to access a
dataset.
}
\description{
Scrapes the dataset information page at
https://www.huduser.gov/portal/datasets/update-schedule.html.
}
