% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_fif}
\alias{hud_fif}
\title{hud_fif}
\usage{
hud_fif(year = 2021)
}
\arguments{
\item{year}{The specific year for the Renewal Funding Inflation Factors.
Supports years 2021 to 2017.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Renewal
Funding Inflation Factors.
}
