% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huduser.R
\name{hud_fmr}
\alias{hud_fmr}
\title{hud_fmr}
\usage{
hud_fmr(
  query,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{query}{Can provide either a 5 digit FIPS code + 99999 at end,
state abbreviation, or CBSA code.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}
}
\value{
This function returns a dataframe containing FAIR MARKETS RENT
data for a particular county or state. For county level data, these
measurements include the county_name, counties_msa, town_name, metro_status,
metro_name, smallarea_status, basicdata, Efficiency, One-Bedroom,
Two-Bedroom, Three-Bedroom, Four-Bedroom, and year For more details about
these measurements, go to https://www.huduser.gov/portal/dataset/fmr-api.html
For state level data, these measurements will be the same as county level
data, but will return a dataframe with the individual measurements for each
individual county within the state.
}
\description{
This function queries the Fair Markets Rent API provided by
US Department of Housing and Urban Development.
}
\keyword{API}
\keyword{Crosswalks}
