% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_fmr.R
\name{hud_fmr_county_zip}
\alias{hud_fmr_county_zip}
\title{Fair Markets Rent County Queries for Zip}
\usage{
hud_fmr_county_zip(
  county,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble = getOption("rhud_use_tibble", FALSE)
)
}
\arguments{
\item{county}{A character or numeric vector: a county to query for.
Must be provided as a 5 digit fipcode.}

\item{year}{A character of numeric vector: gets the year
that this data was recorded.
Can specify multiple year(s). Default is the
previous year.}

\item{key}{A character vector of length one with the key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{A logical: if TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A data frame with fair markets rent for zip codes in counties for
all combinations of "county" and "year" inputs.
}
\description{
This function queries for county(s) and returns FMR calculation.
If the county is not
a small area, it will return only single
measurement for that county. If the county is considered a small area,
it will return data at a zip code level.
}
\examples{
\dontrun{
hud_fmr_county_zip("5100199999", year = c(2021))

hud_fmr_county_zip("5100199999", year = c("2021"))

hud_fmr_county_zip(5151099999, year = c(2021))
}
}
\seealso{
\itemize{
\item \code{\link[=hud_fmr_state_metroareas]{hud_fmr_state_metroareas()}}
\item \code{\link[=hud_fmr_state_counties]{hud_fmr_state_counties()}}
\item \code{\link[=hud_fmr_metroarea_zip]{hud_fmr_metroarea_zip()}}
\item \code{\link[=hud_fmr_county_zip]{hud_fmr_county_zip()}}
\item \code{\link[=hud_fmr]{hud_fmr()}}
}
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
