% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_key.R
\name{hud_get_key}
\alias{hud_get_key}
\title{hud_get_key}
\usage{
hud_get_key()
}
\value{
Returns a string.
}
\description{
Return most recent key set in the HUD_KEY environment variable.
If no key is set, return "".
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_get_key()
}
}
\seealso{
\itemize{
\item \code{\link[=hud_get_key]{hud_get_key()}}
\item \code{\link[=hud_set_key]{hud_set_key()}}
}
}
