% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_set_user_agent.R
\name{hud_get_user_agent}
\alias{hud_get_user_agent}
\title{Get RHUD User Agent}
\usage{
hud_get_user_agent()
}
\value{
A character vector with the user agent used for querying HUD User
APIs.
}
\description{
Get the most recent user agent set.
}
\examples{
\dontrun{

hud_get_user_agent()

}
}
\seealso{
\itemize{
\item \code{\link[=hud_get_user_agent]{hud_get_user_agent()}}
\item \code{\link[=hud_set_user_agent]{hud_set_user_agent()}}
}
}
