% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_hads}
\alias{hud_hads}
\title{hud_hads}
\usage{
hud_hads(year = 2013)
}
\arguments{
\item{year}{The specific year for the Housing Affordability Data System to
query for.
Years are recorded every two years (2013,2011,2009).
No data past 2013 available.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the
Housing Affordability Data System.
}
