% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_hic}
\alias{hud_hic}
\title{hud_hic}
\usage{
hud_hic(year = 2020, res = "s")
}
\arguments{
\item{year}{The specific year for the Housing Inventory Count (HIC) count
to query for. Supports years 2020 to 2007.}

\item{res}{The resolution of the dataset. Either Coc or State level data.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the
Housing Inventory Count (HIC).
}
