% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_hvl}
\alias{hud_hvl}
\title{hud_hvl}
\usage{
hud_hvl(year = 2021)
}
\arguments{
\item{year}{The specific year for the HOME Investment Partnerships
Program (HOME) and Housing Trust Fund (HTF) count to query for.
Supports years 2021 to 2014.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the
HOME Investment Partnerships Program (HOME) and Housing Trust Fund (HTF).
}
