% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huduser.R
\name{hud_il}
\alias{hud_il}
\title{hud_il}
\usage{
hud_il(query, year = 2021, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{query}{Can provide either a 10 digit FIPS code including county
subdivision, or state abbreviation.}

\item{year}{Gets the year that this data was recorded.
Default is the latest year.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
This function returns a dataframe containing INCOME LIMITS data
for a particular county or state. For county level data, these measurements
include the county_name, counties_msa, town_name, metro_status, metro_name,
year, median_income, very_low+, extremely_low+, and low+. For more details
about these measurements, go to
https://www.huduser.gov/portal/dataset/fmr-api.html
}
\description{
This function queries the Income Limits API provided by
US Department of Housing and Urban Development
}
\keyword{API}
\keyword{Income}
\keyword{Limits}
