% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_lithtc}
\alias{hud_lithtc}
\alias{hud_lihtc}
\title{hud_lithtc}
\usage{
hud_lihtc(file = 1)
}
\arguments{
\item{file}{The specific file needed.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from https://lihtc.huduser.gov/
}
