% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_minor_civil_divisions}
\alias{hud_minor_civil_divisions}
\title{hud_minor_civil_divisions}
\usage{
hud_minor_civil_divisions(state)
}
\value{
A dataframe containing details of minor civil divisions in a state
}
\description{
Get a list of all minor civil divisions in a state
}
\keyword{CBSA}
