% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_msil}
\alias{hud_msil}
\title{hud_msil}
\usage{
hud_msil(year = 2021, avg = FALSE)
}
\arguments{
\item{year}{The year to query for in Multifamily Tax Subsidy Projects (MTSP)
Income Limits. Supports years 2021 to 2014.}

\item{avg}{The averaged or non averaged dataset.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the
Multifamily Tax Subsidy Projects (MTSP) Income Limits.
}
