% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_muaf}
\alias{hud_muaf}
\title{hud_muaf}
\usage{
hud_muaf(year = 2022)
}
\arguments{
\item{year}{The specific year for the Utility Allowance Factors (UAFs) for
projects. Supports years 2022 to 2015.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Utility
Allowance Factors (UAFs) for projects.
}
