% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_nation_states_territories}
\alias{hud_nation_states_territories}
\title{hud_nation_states_territories}
\usage{
hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}
}
\value{
A dataframe containing details of all the states and territories
in the US.
}
\description{
Get a list of state and US territories
along with the corresponding fips code and
abbreviation.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_nation_states_territories()
}
}
\keyword{States}
