% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hud_nsp}
\alias{hud_nsp}
\title{hud_nsp}
\usage{
hud_nsp(file = 1)
}
\arguments{
\item{file}{The specific file needed.}
}
\value{
A data frame built from file.
}
\description{
Returns a data frame pertaining to the Neighborhood
Stabilization Program provided by HUD User or HUD Exchange
}
