% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_find_most_recent_data.R
\name{hud_rec_fmr_yr}
\alias{hud_rec_fmr_yr}
\title{Get Most Recently Released Year: Fair Markets Rent}
\usage{
hud_rec_fmr_yr(key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}
}
\value{
The most recent year for the fmr files for state, county, and
metroarea queries.
}
\description{
Pings the Fair Markets Rent API provided by
(US Department of Housing and Urban Development) HUD User to
determine the most recently released files. This will only ping
for the last two years. Will return years for state, county, and metroarea
resolutions.
}
\examples{
\dontrun{

hud_rec_fmr_yr()

}
}
\seealso{
\itemize{
\item \code{\link[=hud_rec_cw_yr]{hud_rec_cw_yr()}}
\item \code{\link[=hud_rec_fmr_yr]{hud_rec_fmr_yr()}}
\item \code{\link[=hud_rec_il_yr]{hud_rec_il_yr()}}
}
}
