% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudkey.R
\name{hud_set_key}
\alias{hud_set_key}
\title{hud_set_key}
\usage{
hud_set_key(key)
}
\arguments{
\item{key}{key obtained at
https://www.huduser.gov/hudapi/public/register?comingfrom=1}
}
\description{
A wrapper around Sys.getenv() to set HUD_KEY environment
variable.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
hud_get_key()

hud_set_key("dwqkqodkqodqkqoqdq")
hud_get_key()
}
}
