% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_state_counties}
\alias{hud_state_counties}
\title{hud_state_counties}
\usage{
hud_state_counties(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all counties.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing all counties within a state
}
\description{
Get a list of all counties within a state.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_counties("CA")
hud_state_counties("Virginia")
hud_state_counties("51")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{Counties}
