% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_state_metropolitan}
\alias{hud_state_metropolitan}
\title{hud_state_metropolitan}
\usage{
hud_state_metropolitan(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all the metropolitan areas.}

\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}
}
\value{
A dataframe containing details of metropolitan areas in US.
}
\description{
Get a list of all metropolitan areas in the US along with its
name and CBSA code.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_metropolitan("VA")
}
}
\keyword{CBSA}
