% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_misc.R
\name{hud_state_minor_civil_divisions}
\alias{hud_state_minor_civil_divisions}
\title{US Minor Civil Divisions}
\usage{
hud_state_minor_civil_divisions(state, key = Sys.getenv("HUD_KEY"), to_tibble)
}
\arguments{
\item{state}{The state to get all mcd(s).}

\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing details of minor civil divisions in state(s).
}
\description{
Get a list of all minor civil divisions in a state
}
\examples{
\dontrun{
hud_state_minor_civil_divisions("CA")
hud_state_minor_civil_divisions("Virginia")
hud_state_minor_civil_divisions("51")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{mcd}
