% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudexchange.R
\name{hudcdbg}
\alias{hudcdbg}
\title{hudcdbg}
\usage{
hudcdbg(file = 1)
}
\arguments{
\item{file}{The specific file needed. Currently you can specify 1 -> 4. Will default to 1:
\enumerate{
\item DR CDBG CDBG-DR Financial Report by Appropriation (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Financial Report by Grantee (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Financial Report Monthly Summary (xlsx - 0.1 MB - 1/1/2022)
\item DR CDBG CDBG-DR Performance by Activity (xlsx - 2.2 MB - 1/1/2022)
}}
}
\value{
A dataframe.
}
\description{
This will grab data from https://drgr.hud.gov/public/data_downloads.html?programName=DR\%20CDBG
and format it into a dataframe. It will unroll each field to remove NA values in cells.
}
