% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudcinch}
\alias{hudcinch}
\title{hudcinch}
\usage{
hudcinch(year = "2017-2015")
}
\arguments{
\item{year}{The specific year for the Components of Inventory Change (CINCH) report to query for. Years must be a character type specifying
a year range:
\enumerate{
\item 2017-2015,
\item 2011-2013,
\item 2009-2011,
\item 2009 -2007,
\item 2003-2005,
\item 2001-2003
}}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Components of Inventory Change (CINCH) report.
}
