% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{huddatasets}
\alias{huddatasets}
\title{huddatasets}
\usage{
huddatasets()
}
\value{
Dataframe with information about the various datasets that HUD has, update information, and corresponding functions and documentation to access a dataset using the package.
}
\description{
Scrapes the dataset information page at https://www.huduser.gov/portal/datasets/update-schedule.html.
}
\details{
huddatasets
}
