% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{huddda}
\alias{huddda}
\title{huddda}
\usage{
huddda(year = 2022, sheet = 1)
}
\arguments{
\item{year}{The specific year for the Difficult Development Areas to query for.}

\item{sheet}{The specific sheet in the file.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Difficult Development Areas.
}
