% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huduser.R
\name{hudfmr}
\alias{hudfmr}
\title{hudfmr}
\usage{
hudfmr(query, year = pkg.env$curr.year, key = pkg.env$curr.key)
}
\arguments{
\item{query}{Can provide either a 10 digit FIPS code including county subdivision, or state abbreviation.}

\item{year}{Gets the year that this data was recorded. Default is the latest year.}

\item{key}{The API key for this user. You must go to HUD and sign up for an account and request for an API key.}
}
\value{
This function returns a dataframe containing FAIR MARKETS RENT data for a particular county or state.
For county level data, these measurements include the county_name, counties_msa, town_name, metro_status, metro_name,
smallarea_status, basicdata, Efficiency, One-Bedroom, Two-Bedroom, Three-Bedroom, Four-Bedroom, and year
For more details about these measurements, go to https://www.huduser.gov/portal/dataset/fmr-api.html
For state level data, these measurements will be the same as county level data, but will return a dataframe with
the individual measurements for each individual county within the state.
}
\description{
This function queries the Fair Markets Rent API provided by US Department of Housing and Urban Development
}
\keyword{API}
\keyword{Crosswalks}
