% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{hudgetkey}
\alias{hudgetkey}
\title{hudgetkey}
\usage{
hudgetkey()
}
\value{
Will return the most recent key set. If no key is set it will return NULL.
}
\description{
Will return the most recent key set. If no key is set it will return NULL.
}
\details{
hudgetkey
}
