% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudhaf}
\alias{hudhaf}
\title{hudhaf}
\usage{
hudhaf(year = 2022)
}
\arguments{
\item{year}{The specific year for the Homeowner Assistance Fund Income Limits to queryt for.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Homeowner Assistance Fund Income Limits
}
