% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudlithtc}
\alias{hudlithtc}
\alias{hudlihtc}
\title{hudlithtc}
\usage{
hudlihtc(file = 1)
}
\arguments{
\item{file}{The specific file needed.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from https://lihtc.huduser.gov/
}
