% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudpit}
\alias{hudpit}
\title{hudpit}
\usage{
hudpit(year = 2020, res = "s", pop = "s")
}
\arguments{
\item{year}{The specific year for the Point-In-Time (PIT) count to query for. Supports years 2020 to 2007.}

\item{res}{The resolution of the dataset. Either Coc or State level data.}

\item{pop}{Either v for veterans or s for standard population.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Point-In-Time (PIT) count.
}
