% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudwebsitedocs.R
\name{hudr_website}
\alias{hudr_website}
\title{hudr_website}
\usage{
hudr_website(website = c("github-pages", "github"))
}
\arguments{
\item{website}{The websites available.
\enumerate{
\item "github-pages"
\item "github"
}}
}
\description{
Quickly get documentation for the hudr package by opening up
the websites associated with it. Currently supports Unix and Windows OS.
}
