% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudrfif}
\alias{hudrfif}
\title{hudrfif}
\usage{
hudrfif(year = 2021)
}
\arguments{
\item{year}{The specific year for the Renewal Funding Inflation Factors. Supports years 2021 to 2017.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Renewal Funding Inflation Factors.
}
