% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{hudsetkey}
\alias{hudsetkey}
\title{hudsetkey}
\usage{
hudsetkey(key)
}
\arguments{
\item{key}{The token given by USPS}
}
\description{
The function will save the key into your library download. You will need to
update this for new library(hudr) call. You can also set a new key by using this function with a new one.
}
\details{
hudsetkey
}
