% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudusernonapi.R
\name{hudsfmr}
\alias{hudsfmr}
\title{hudsfmr}
\usage{
hudsfmr(year = 2022)
}
\arguments{
\item{year}{The specific year for the Small Area Fair Market Rents. Supports years 2022 to 2016.}
}
\value{
A data frame built from the file.
}
\description{
Returns a data frame from HUD USER pertaining to the Small Area Fair Market Rents (SAFMRs).
}
