% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_check_internet.R
\name{is_internet_available}
\alias{is_internet_available}
\title{Check if Internet is Available}
\usage{
is_internet_available()
}
\value{
TRUE if internet is available, halts execution if not.
}
\description{
A wrapper around curl::has_internet() to check
if internet is available.
}
