% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{misc_do_query_call}
\alias{misc_do_query_call}
\title{misc_do_query_call}
\usage{
misc_do_query_call(URL, key)
}
\arguments{
\item{URL}{The URLs to query for.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing all queried rows.
}
\description{
Make queries calls given a list of URLs
}
