% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudnumbersonly.R
\name{numbers_only}
\alias{numbers_only}
\title{numbers_only}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{A string.}
}
\value{
TRUE if string is all numbers, FALSE if not.
}
\description{
Detect whether string contains only numbers.
}
