% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudparallel.R
\name{parallelize_api_calls}
\alias{parallelize_api_calls}
\title{parallelize_api_calls}
\usage{
parallelize_api_calls(allqueries, cores = 1)
}
\arguments{
\item{cores}{The number of CPU cores to use when processing the requests.}

\item{urls}{A list of urls to query for.}
}
\value{
The dataframe collection of all the response bodies.
}
\description{
Make API calls using multiple cores.
}
\keyword{parallel}
