% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_cache.R
\name{rhud_clear_cache}
\alias{rhud_clear_cache}
\title{rhud_clear_cache}
\usage{
rhud_clear_cache()
}
\description{
Remove cached data from the caching directory that is used to
store data retrieved using the rhud API calls. By default, rhud uses a
non-persistent temporary directory given for an R session, but a user might
have set another directory to use.
}
\examples{
\dontrun{
rhud_clear_cache()

}
}
