% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhud_sitrep.R
\name{rhud_sitrep}
\alias{rhud_sitrep}
\title{Package Status}
\usage{
rhud_sitrep()
}
\description{
Checks whether the rhud key has been set, if rhud can
query some the APIs, and displays the current cache settings
for the package.
}
\examples{
\dontrun{

rhud_sitrep()

}
}
