% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{setkey}
\alias{setkey}
\title{setkey}
\usage{
setkey(key)
}
\arguments{
\item{key}{The token given by USPS}
}
\description{
The function will save the key into your package download. You will need to
update this for new downloads. You can also set a new key by using this function with a new one.
}
\details{
setkey
}
