% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_in_geo.R
\name{z_in_cd}
\alias{z_in_cd}
\title{z_in_cd}
\usage{
z_in_cd(zip, cd, year, quarter)
}
\arguments{
\item{zip}{The zip to determine overlap with cd}

\item{cd}{The cd to determine overlap with zip}

\item{year}{The year of the crosswalk files.}

\item{quarter}{The quarter of the crosswalk files.}
}
\value{
If zip(s) exist in the cd(s) specified, then TRUE is returned.
}
\description{
Given a zip code and a congressional district, determine
if they overlap using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

z_in_cd(zip = 35213, cd = 2204, year = 2019, quarter = 2)

}
}
\seealso{
\itemize{
\item \code{\link[=z_in_trt]{z_in_trt()}}
\item \code{\link[=z_in_cty]{z_in_cty()}}
\item \code{\link[=z_in_cbsa]{z_in_cbsa()}}
\item \code{\link[=z_in_cbsadiv]{z_in_cbsadiv()}}
\item \code{\link[=z_in_ctysb]{z_in_ctysb()}}
\item \link{z_in_cd}
\item \code{\link[=trt_in_z]{trt_in_z()}}
\item \code{\link[=cty_in_z]{cty_in_z()}}
\item \code{\link[=cbsa_in_z]{cbsa_in_z()}}
\item \code{\link[=cbsadiv_in_z]{cbsadiv_in_z()}}
\item \code{\link[=cd_in_z]{cd_in_z()}}
\item \code{\link[=ctysb_in_z]{ctysb_in_z()}}
}
}
