% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfixfips.R
\name{add_leading_zeros}
\alias{add_leading_zeros}
\title{add_leading_zeros}
\usage{
add_leading_zeros(geoid_type = "zip", geoids, num_chars = -Inf)
}
\arguments{
\item{geoid_type}{What geoid type to add leading 0s to
\enumerate{
\item zip -> Must be 5 digit zip code
\item county -> Must be a 5 digit county fips code
\item cbsa -> Must be a 5 digit cbsa code
\item cbsa-div -> Must be a 5 digit cbsa code
\item countysub -> Must be a 10 digit countysub code
\item cd -> Must be a 4 digit cd code
}}

\item{geoids}{A vector or column of geoids that need leading zeros to be processed properly.}

\item{num_chars}{If you don't know what geoid_type you are looking at, you can specify the length this geoid should be.}
}
\value{
A dataframe with the corrected geoids.
}
\description{
Sometimes when loading in fips data from an excel file it truncates the leading 0's. This attempts to re-add those.
}
