% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudaddleadingzeros.R
\name{add_leading_zeros}
\alias{add_leading_zeros}
\title{add_leading_zeros}
\usage{
add_leading_zeros(geoid_type = "zip", geoids)
}
\arguments{
\item{geoid_type}{What geoid type to add leading 0s to 1) zip -> Must be 5
digit zip code 2) county -> Must be a 5 digit county fips code 3) cbsa ->
Must be a 5 digit cbsa code 5) countysub -> Must be a 10 digit countysub
code 6) cd -> Must be a 4 digit cd code. If you specify a number, it will
make the geoid of that length.}

\item{geoids}{A vector or column of geoids that need leading zeros to be
processed properly.}
}
\value{
A dataframe with the corrected geoids.
}
\description{
Sometimes when loading in data from an excel file it
truncates the leading 0's. This attempts to re-add those.
}
\examples{
\dontrun{
library(hudr)
zip <- c(02102, 11032, 01232)

zip <- add_leading_zeros("zip", zip)

cbsa <- c(02102, 11032, 01232)

cbsa <- add_leading_zeros("cbsa", cbsa)
}
}
