% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huddatasetcw.R
\name{crosswalk}
\alias{crosswalk}
\title{crosswalk}
\usage{
crosswalk(
  data,
  geoid,
  geoid_col,
  cw_geoid,
  cw_geoid_col = NA,
  method = NA,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{data}{A dataset with rows describing measurements at a zip,
county, county subdivision (countysub), congressional district (cd),
census tract, core base statistical area (cbsa), or core based
statistical area division (cbsadiv) geoid.
\enumerate{
\item zip
\item tract
\item county
\item countysub
\item cbsa
\item cbsadiv
\item cd
}}

\item{geoid}{The current geoid that the dataset is described in: must be
zip, county, countysub, cd,
tract, cbsa, or cbsadiv geographic id.
\enumerate{
\item zip
\item tract
\item county
\item countysub
\item cbsa
\item cbsadiv
\item cd
}}

\item{geoid_col}{The column containing the geographic identifier; must be
zip, county, county subdivision (countysub), congressional district (cd),
census tract, core base statistical area (cbsa), and core based
statistical area division (cbsadiv) geoid.
Supply either the name of the column or the index.
All elements in this column must be numbers only at the proper length.
For example, zip codes must be 5 digit numbers.}

\item{cw_geoid}{The geoid to crosswalk the dataset to; must be
zip, county, county subdivision (countysub), congressional district (cd),
census tract, core base statistical area (cbsa), or core based
statistical area division (cbsadiv) geoid.
\enumerate{
\item zip
\item tract
\item county
\item countysub
\item cbsa
\item cbsadiv
\item cd
}}

\item{cw_geoid_col}{The columns in the dataset to distribute
according to method ratio.
If method is empty, no allocation method will be applied --
the crosswalk file will just be merged to the dataset.
All elements in these columns must be numbers only.}

\item{method}{The allocation method to use: residential,
business, other, or total. If method is empty, no allocation
method will be applied -- the crosswalk file will just be merged
to the dataset.
\enumerate{
\item res
\item bus
\item tot
\item oth
}}

\item{year}{The year measurement was taken.}

\item{quarter}{The quarter of year measurement was taken.}

\item{key}{The key obtain from HUD USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing the crosswalked dataset.
}
\description{
Using the US Housing and Urban Development crosswalk files,
crosswalk an entire dataset.
Currently supported crosswalks:
\enumerate{
\item zip-tract
\item zip-county
\item zip-cbsa
\item zip-cbsadiv (Available 4th Quarter 2017 onwards)
\item zip-cd
\item tract-zip
\item county-zip
\item cbsa-zip
\item cbsadiv-zip (Available 4th Quarter 2017 onwards)
\item cd-zip
\item zip-countysub (Available 2nd Quarter 2018 onwards)
\item countysub-zip (Available 2nd Quarter 2018 onwards)
}
}
\examples{
\dontrun{
library(rhud)

sample <- data.frame(population = c(42134, 12413, 13132),
                     county = c(24047, 24045, 24043))

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "res")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "bus")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "bus",
          year = 2018, quarter = 1)

}
}
\seealso{
\itemize{
\item \code{\link[=crosswalk]{crosswalk()}}
\item \code{\link[=hud_cw_zip_tract]{hud_cw_zip_tract()}}
\item \code{\link[=hud_cw_zip_county]{hud_cw_zip_county()}}
\item \code{\link[=hud_cw_zip_cbsa]{hud_cw_zip_cbsa()}}
\item \code{\link[=hud_cw_zip_cbsadiv]{hud_cw_zip_cbsadiv()}}
\item \code{\link[=hud_cw_zip_countysub]{hud_cw_zip_countysub()}}
\item \code{\link[=hud_cw_zip_cd]{hud_cw_zip_cd()}}
\item \code{\link[=hud_cw_tract_zip]{hud_cw_tract_zip()}}
\item \code{\link[=hud_cw_county_zip]{hud_cw_county_zip()}}
\item \code{\link[=hud_cw_cbsa_zip]{hud_cw_cbsa_zip()}}
\item \code{\link[=hud_cw_cbsadiv_zip]{hud_cw_cbsadiv_zip()}}
\item \code{\link[=hud_cw_cd_zip]{hud_cw_cd_zip()}}
\item \code{\link[=hud_cw_countysub_zip]{hud_cw_countysub_zip()}}
\item \code{\link[=hud_cw]{hud_cw()}}
}
}
