% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huddatasetcw.R
\name{crosswalk}
\alias{crosswalk}
\title{crosswalk}
\usage{
crosswalk(
  data,
  geoid,
  geoid_col,
  cw_geoid,
  cw_geoid_col = NA,
  method = NA,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{data}{A dataset with rows describing measurements at a zip,
county, countysub, cd,
tract, cbsa, or cbsadiv geographic id.}

\item{geoid}{The current geoid that the dataset is described in: must be
zip, county, countysub, cd,
tract, cbsa, or cbsadiv geographic id.}

\item{geoid_col}{The column containing the geographic identifier; must be
zip, county, countysub, cd,
tract, cbsa, or cbsadiv geographic id.
Supply either the name of the column or the index.}

\item{cw_geoid}{The geoid to crosswalk the dataset to.}

\item{cw_geoid_col}{The columns in the dataset to distribute
according to method ratio.
If method is empty, no allocation method will be applied --
the crosswalk file will just be attached to the dataset.}

\item{method}{The allocation method to use: residential, business, other, or total.
If method is empty, no allocation method will be applied --
the crosswalk file will just be attached to the dataset.}

\item{year}{The year measurement was taken.}

\item{quarter}{The quarter of year measurement was taken.}

\item{key}{The key obtain from HUD USER website.}
}
\value{
A dataframe containing the crosswalked dataset.
}
\description{
Using the crosswalk files, crosswalk an entire dataset.
Currently supported crosswalks:
\enumerate{
\item zip-tract
\item zip-county
\item zip-cbsa
\item zip-cbsadiv (Available 4th Quarter 2017 onwards)
\item zip-cd
\item tract-zip
\item county-zip
\item cbsa-zip
\item cbsadiv-zip (Available 4th Quarter 2017 onwards)
\item cd-zip
\item zip-countysub (Available 2nd Quarter 2018 onwards)
\item countysub-zip (Available 2nd Quarter 2018 onwards)
}
}
