% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%ctysb_in_z\%}
\alias{\%ctysb_in_z\%}
\title{\%ctysb_in_z\%}
\usage{
countysub \%ctysb_in_z\% zip
}
\arguments{
\item{countysub}{The countysub(s) to determine overlap with zip(s).}

\item{zip}{The zip(s). to determine overlap with countysub(s).}
}
\value{
If countysub(s) exist in the zip(s) specified, then TRUE is returned.
}
\description{
To use this function, the HUD_KEY system environment
variable must be set first: use hud_set_key("the_key") to do this.

Given countysub(s) and zip(s), determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted
even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.

Infix operators default to most recent year and quarter of which the
crosswalk files are available. For modifying the year and quarter
parameter,
take a look at their non-infix versions.
}
\examples{
\dontrun{

2203194756 \%ctysb_in_z\% 71052

}
}
