% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%ctysb_in_z\%}
\alias{\%ctysb_in_z\%}
\title{\%ctysb_in_z\%}
\usage{
countysub \%ctysb_in_z\% zip
}
\arguments{
\item{countysub}{The countysub to determine overlap with zip}

\item{zip}{The zip to determine overlap with countysub}
}
\value{
If countysub(s) exist in the zip(s) specified, then TRUE is returned.
}
\description{
Given a countysub and zip code, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.

Infix operators default to most recent year and quarter of which the
crosswalk files are available. For access to the year and quarter parameter,
take a look at their non-infix versions.
}
\examples{
\dontrun{

2203194756 \%ctysb_in_z\% 71052

}
}
