% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%trt_in_z\%}
\alias{\%trt_in_z\%}
\title{\%trt_in_z\%}
\usage{
tract \%trt_in_z\% zip
}
\arguments{
\item{tract}{The tract to determine overlap with zip}

\item{zip}{The zip to determine overlap with tract}
}
\description{
Given a tract and a zip code, determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.
}
\examples{
\dontrun{

22031950600 \%trt_in_z\% 71052

}
}
