% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_cd\%}
\alias{\%z_in_cd\%}
\title{\%z_in_cd\%}
\usage{
zip \%z_in_cd\% cd
}
\arguments{
\item{zip}{The zip(s) to determine overlap with a congressional district(s)}

\item{cd}{The congressional district(s) to determine overlap with zip(s)}
}
\value{
If zip(s) exist in the cd(s) specified, then TRUE is returned.
}
\description{
To use this function, the HUD_KEY system environment
variable must be set first: use hud_set_key("the_key") to do this.

Given zip code(s) and congressional district(s), determine if
they overlap using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.

Infix operators default to most recent year and quarter of which the
crosswalk files are available. For modifying the year and quarter parameter,
take a look at their non-infix versions.
}
\examples{
\dontrun{

71052 \%z_in_cd\% 2204

}
}
