% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geo_is_infix_geo.R
\name{\%z_in_ctysb\%}
\alias{\%z_in_ctysb\%}
\title{\%z_in_ctysb\%}
\usage{
zip \%z_in_ctysb\% countysub
}
\arguments{
\item{zip}{The zip(s) to determine overlap with countysub(s)}

\item{countysub}{The countysub(s) to determine overlap with zip(s)}
}
\value{
If zip(s) exist in the countysub(s) specified, then TRUE is returned.
}
\description{
To use this function, the HUD_KEY system environment
variable must be set first: use hud_set_key("the_key") to do this.

Given zip code(s) and a countysub(s), determine if they overlap
using the crosswalk files. Overlap will be described if
any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted
even
though their boundaries intersect the queried geoids(rhs). This is likely
because addresses do not lie in their intersecting region.

Infix operators default to most recent year and quarter of which the
crosswalk files are available. For modifying the year and quarter
parameter,
take a look at their non-infix versions.
}
\examples{
\dontrun{

71052 \%z_in_ctysb\% 2203194756

}
}
