% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_omni_functions.R
\name{hud_chas}
\alias{hud_chas}
\title{hud_chas}
\usage{
hud_chas(
  type,
  state_id = NULL,
  entity_id = NULL,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{type}{Queries the data based off:
1 - Nation
2 - State
3 - County
4 - MCD
5 - Place}

\item{state_id}{For types 2,3,4,5, you must provide a state_id.}

\item{entity_id}{For types 3,4,5, you must provide a fips code, placecode,
or mcd code.}

\item{year}{Gets the year that this data was recorded. Defaults to 2014-2018.
There are specific year ranges that are only accepted.
\itemize{
\item 2014-2018
\item 2013-2017
\item 2012-2016
\item 2011-2015
\item 2010-2014
\item 2009-2013
\item 2008-2012
\item 2007-2011
\item 2006-2010
}}

\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
This function returns a dataframe or tibble containing
Comprehensive Housing
and Affordability Strategy data for a particular state.
For more details about
these measurements, go to
https://www.huduser.gov/portal/dataset/chas-api.html
}
\description{
This function queries the
Comprehensive Housing Affordability Strategy CHAS API provided by
US Department
of Housing and Urban Development (HUD USER). The ordering of items in
stateid input must match that of the entityid input.
}
\examples{
\dontrun{
hud_chas(1)

hud_chas('2', state_id = '56')

hud_chas('3','51','199')

hud_chas('4', '51', 94087)

hud_chas('5', '51', 48996)
}
}
\keyword{(CHAS)}
\keyword{API}
\keyword{Affordability}
\keyword{Comprehensive}
\keyword{Housing}
\keyword{Strategy}
