% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_chas.R
\name{hud_chas_county}
\alias{hud_chas_county}
\title{Comprehensive Housing and Affordability Strategy for the US Counties}
\usage{
hud_chas_county(
  county,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{county}{The county(s) to query for. Must supply a geoid. 2 digit state fip
\itemize{
\item 3 digit county fip. hud_state_counties() will show an extra 99999 at the
end. Just remove that.
}}

\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
Returns a data frame with
Comprehensive Housing and Affordability Strategy (CHAS) data for
counties for all combinations of "county" and "year" inputs.
}
\description{
Returns Comprehensive Housing and Affordability Strategy (CHAS)
}
\examples{
\dontrun{

hud_chas_county(county = c("06105", "06113"))

}
}
\seealso{
\itemize{
\item \code{\link[=hud_chas_nation]{hud_chas_nation()}}
\item \code{\link[=hud_chas_state]{hud_chas_state()}}
\item \code{\link[=hud_chas_county]{hud_chas_county()}}
\item \code{\link[=hud_chas_state_mcd]{hud_chas_state_mcd()}}
\item \code{\link[=hud_chas_state_place]{hud_chas_state_place()}}
}
}
