% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudchas.R
\name{hud_chas_place}
\alias{hud_chas_place}
\title{hud_chas_place}
\usage{
hud_chas_place(
  state,
  place,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{state}{The state name, abbreviation, or fips code. Make sure if state fips is 1 digit number, do not include leading 0.}

\item{place}{The place(Usually just cities) to query for. Must supply as a 5 digit geoid.}

\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtain from HUD USER website.}
}
\value{
Returns a dataframe with CHAS data for a particular place (Usually just cities).
}
\description{
Returns CHAS data for place.
}
