% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudchas.R
\name{hud_chas_state}
\alias{hud_chas_state}
\title{hud_chas_state}
\usage{
hud_chas_state(state, year = c("2014-2018"), key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to query for. Can supply as abbreviation, whole name,
or as geoid.}

\item{year}{The years to query for.
\itemize{
\item year = "2014-2018"
\item year = "2013-2017"
\item year = "2012-2016"
\item year = "2011-2015"
\item year = "2010-2014"
\item year = "2009-2013"
\item year = "2008-2012"
\item year = "2007-2011"
\item year = "2006-2010"
}}

\item{key}{The key obtain from HUD USER website.}
}
\value{
Returns a dataframe with CHAS data for a particular state.
}
\description{
Returns CHAS data for a state.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_chas_state("CA")
hud_chas_state("New York")
hud_chas_state("51")
}
}
