% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_cities}
\alias{hud_cities}
\title{hud_cities}
\usage{
hud_cities(state)
}
\value{
A dataframe containing details of cities in a state
}
\description{
Get a list of all cities in a state
}
\keyword{Cities}
