% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_counties}
\alias{hud_counties}
\title{hud_counties}
\usage{
hud_counties(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all counties.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing all counties within a state
}
\description{
Get a list of all counties within a state.
}
\keyword{Counties}
