% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_crosswalk.R
\name{hud_cw_countysub_zip}
\alias{hud_cw_countysub_zip}
\title{Crosswalk Countysub to Zip}
\usage{
hud_cw_countysub_zip(
  countysub,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY"),
  to_tibble = getOption("rhud_use_tibble", FALSE)
)
}
\arguments{
\item{countysub}{A character or numeric vector:
10 digit county sub division code.}

\item{year}{A character or numeric vector: gets the year that this data
was recorded. Can specify multiple
years. Default is the previous year.}

\item{quarter}{A character or numeric vector: gets the quarter of the year
that this data was recorded.
Defaults to the first quarter of the year.}

\item{minimal}{A logical: return just the crosswalked geoids if TRUE.
Otherwise, return
all fields. This does not remove duplicates.}

\item{key}{A character vector of length one with the key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}

\item{tract}{A character or numeric vector: 11 digit tract code.}
}
\value{
This function returns a dataframe containing crosswalk data for
countysub to zip for all combinations of "countysub", "year", and "quarter"
inputs.

These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the USPS Crosswalks API provided by US
Department of Housing and Urban Development (HUD USER).
This returns the crosswalk for
countysub to zip.
}
\examples{
\dontrun{

hud_cw_countysub_zip(countysub = '4606720300',
   year = c('2019', '2019', '2019'), quarter = c('4','4'))

hud_cw_countysub_zip(countysub = '3706794068',
   year = c('2019', '2019', '2019'), quarter = c('4','4'))

hud_cw_countysub_zip(countysub = '3711993268',
   year = c('2019', '2019', '2019'), quarter = c('4','4'))

hud_cw_countysub_zip(countysub = '3910383426',
   year = c('2019'), quarter = c('1'), minimal = TRUE)
}
}
\seealso{
\itemize{
\item \code{\link[=hud_cw_zip_tract]{hud_cw_zip_tract()}}
\item \code{\link[=hud_cw_zip_county]{hud_cw_zip_county()}}
\item \code{\link[=hud_cw_zip_cbsa]{hud_cw_zip_cbsa()}}
\item \code{\link[=hud_cw_zip_cbsadiv]{hud_cw_zip_cbsadiv()}}
\item \code{\link[=hud_cw_zip_countysub]{hud_cw_zip_countysub()}}
\item \code{\link[=hud_cw_zip_cd]{hud_cw_zip_cd()}}
\item \code{\link[=hud_cw_tract_zip]{hud_cw_tract_zip()}}
\item \code{\link[=hud_cw_county_zip]{hud_cw_county_zip()}}
\item \code{\link[=hud_cw_cbsa_zip]{hud_cw_cbsa_zip()}}
\item \code{\link[=hud_cw_cbsadiv_zip]{hud_cw_cbsadiv_zip()}}
\item \code{\link[=hud_cw_cd_zip]{hud_cw_cd_zip()}}
\item \code{\link[=hud_cw_countysub_zip]{hud_cw_countysub_zip()}}
\item \code{\link[=hud_cw]{hud_cw()}}
}
}
\keyword{API}
\keyword{Crosswalks}
