% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudcw.R
\name{hud_cw_countysub_zip}
\alias{hud_cw_countysub_zip}
\title{hud_cw_countysub_zip}
\usage{
hud_cw_countysub_zip(
  countysub,
  year = format(Sys.Date() - 365, "\%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{countysub}{10-digit GEOID for the County sub Eg: 4606720300 for type 12}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{quarter}{Gets the quarter of the year that this data was recorded.
Defaults to the first quarter of the year.}

\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return
all fields. This does not remove duplicates.}

\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}
}
\value{
This function returns a dataframe containing CROSSWALK data for a
particular GEOID. These measurements include res-ratio, bus-ratio,
oth-ratio, tot-ratio. For more details on these measurements, visit
https://www.huduser.gov/portal/dataset/uspszip-api.html
}
\description{
This function queries the Crosswalks API provided by
US Department of Housing and Urban Development. This
returns the crosswalk for countysub to zip.
(Available 2nd Quarter 2018 onwards)
}
\examples{
\dontrun{
library(hudr)

hud_cw_countysub_zip(countysub = '4606720300 ',
   year = c('2019', '2019', '2019'), quarter = c('4','4'))

hud_cw_countysub_zip(countysub = '4606720300 ',
   year = c('2019'), quarter = c('1'), minimal = TRUE)
}
}
\seealso{
\itemize{
\item \code{\link[=hud_cw_zip_tract]{hud_cw_zip_tract()}}
\item \code{\link[=hud_cw_zip_county]{hud_cw_zip_county()}}
\item \code{\link[=hud_cw_zip_cbsa]{hud_cw_zip_cbsa()}}
\item \code{\link[=hud_cw_zip_cbsadiv]{hud_cw_zip_cbsadiv()}}
\item \code{\link[=hud_cw_zip_countysub]{hud_cw_zip_countysub()}}
\item \code{\link[=hud_cw_zip_cd]{hud_cw_zip_cd()}}
\item \code{\link[=hud_cw_tract_zip]{hud_cw_tract_zip()}}
\item \code{\link[=hud_cw_county_zip]{hud_cw_county_zip()}}
\item \code{\link[=hud_cw_cbsa_zip]{hud_cw_cbsa_zip()}}
\item \code{\link[=hud_cw_cbsadiv_zip]{hud_cw_cbsadiv_zip()}}
\item \code{\link[=hud_cw_cd_zip]{hud_cw_cd_zip()}}
\item \code{\link[=hud_cw_countysub_zip]{hud_cw_countysub_zip()}}
\item \code{\link[=hud_cw]{hud_cw()}}
}
}
\keyword{API}
\keyword{Crosswalks}
