% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfmr.R
\name{hud_fmr_county_zip}
\alias{hud_fmr_county_zip}
\title{hud_fmr_county_zip}
\usage{
hud_fmr_county_zip(
  county,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{county}{A county to query for.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}
}
\value{

}
\description{
This function queries for a county and returns FMR calculation.
If the county is not
a small area, it will return only single
measurement for that county. If the county is considered a small area,
it will return data at a zip code level.
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
