% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_fmr.R
\name{hud_fmr_metroarea_zip}
\alias{hud_fmr_metroarea_zip}
\title{hud_fmr_metroarea_zip}
\usage{
hud_fmr_metroarea_zip(
  metroarea,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{metroarea}{Metroareas to query for.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A data frame with fair markets rent for zip codes in metro areas for
all combinations of "metroarea" and "year" inputs.
}
\description{
This function queries for a metroarea and returns
FMR calculation. If the metroarea is not
a small area, it will return only single
measurement for that metroarea. If the metrarea is considered a
small area, it will return data at a zip code level.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_fmr_metroarea_zip("METRO47900M47900", year = c(2018))

hud_fmr_metroarea_zip("METRO29180N22001", year = c(2019))

hud_fmr_metroarea_zip("METRO10380M10380", year = c(2020))
}
}
\seealso{
\itemize{
\item \code{\link[=hud_fmr_state_metroareas]{hud_fmr_state_metroareas()}}
\item \code{\link[=hud_fmr_state_counties]{hud_fmr_state_counties()}}
\item \code{\link[=hud_fmr_metroarea_zip]{hud_fmr_metroarea_zip()}}
\item \code{\link[=hud_fmr_county_zip]{hud_fmr_county_zip()}}
}
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
