% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudfmr.R
\name{hud_fmr_state_metroareas}
\alias{hud_fmr_state_metroareas}
\title{hud_fmr_state_metroareas}
\usage{
hud_fmr_state_metroareas(
  state,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY")
)
}
\arguments{
\item{state}{The state to query for.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}
}
\value{
Fair markets rent for metroareas.
}
\description{
This function queries for a state and returns the FMR calculation
at a metroarea resolution.
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
