% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_fmr.R
\name{hud_fmr_state_metroareas}
\alias{hud_fmr_state_metroareas}
\title{hud_fmr_state_metroareas}
\usage{
hud_fmr_state_metroareas(
  state,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{state}{The state to query for. Can be abbreviation, fip code, or
full name.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A data frame with fair markets rent for metro areas in states for
all combinations of "state" and "year" inputs.
}
\description{
This function queries for a state and returns the
FMR calculation
at a metroarea resolution for all metroareas in this state.
}
\examples{
\dontrun{

hud_fmr_state_metroareas("VA", year = c(2021))

hud_fmr_state_metroareas("Alabama", year = c(2021))

hud_fmr_state_metroareas("24", year = c(2021))

}
}
\seealso{
\itemize{
\item \code{\link[=hud_fmr_state_metroareas]{hud_fmr_state_metroareas()}}
\item \code{\link[=hud_fmr_state_counties]{hud_fmr_state_counties()}}
\item \code{\link[=hud_fmr_metroarea_zip]{hud_fmr_metroarea_zip()}}
\item \code{\link[=hud_fmr_county_zip]{hud_fmr_county_zip()}}
}
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
