% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_fmr.R
\name{hud_fmr_state_metroareas}
\alias{hud_fmr_state_metroareas}
\title{Fair Markets Rent State Queries for Metroareas}
\usage{
hud_fmr_state_metroareas(
  state,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{state}{The state(s) to query for. Can be abbreviation(s),
fip code(s), or full name(s).}

\item{year}{Gets the year that this data was recorded.
Can specify multiple year(s). Default is the
previous year.}

\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A data frame with Fair Markets Rent for metro areas in state(s) for
all combinations of "state" and "year" inputs.
}
\description{
This function queries for a state and returns the
(Fair Markets Rent) FMR calculation
at a metroarea resolution for all metroareas in the state query.
}
\examples{
\dontrun{

hud_fmr_state_metroareas("VA", year = c(2021))

hud_fmr_state_metroareas("Alabama", year = c(2021))

hud_fmr_state_metroareas("24", year = c(2021))

}
}
\seealso{
\itemize{
\item \code{\link[=hud_fmr_state_metroareas]{hud_fmr_state_metroareas()}}
\item \code{\link[=hud_fmr_state_counties]{hud_fmr_state_counties()}}
\item \code{\link[=hud_fmr_metroarea_zip]{hud_fmr_metroarea_zip()}}
\item \code{\link[=hud_fmr_county_zip]{hud_fmr_county_zip()}}
}
}
\keyword{API}
\keyword{Fair}
\keyword{Markets}
\keyword{Rent}
