% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_omni_functions.R
\name{hud_il}
\alias{hud_il}
\title{hud_il}
\usage{
hud_il(
  query,
  year = format(Sys.Date() - 365, "\%Y"),
  key = Sys.getenv("HUD_KEY"),
  to_tibble
)
}
\arguments{
\item{query}{Querying a
state provides the IL measurement for that state level resolution.
Querying a county or cbsa will provide data at a county and
cbsa resolution, respectively.}

\item{year}{Gets the year that this data was recorded.
Can specify multiple years. Default is the
previous year.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
This function returns a dataframe containing INCOME LIMITS data
for a particular county or state. For county level data, these measurements
include the county_name, counties_msa, town_name, metro_status, metro_name,
year, median_income, very_low+, extremely_low+, and low+. For more details
about these measurements, go to
https://www.huduser.gov/portal/dataset/fmr-api.html
}
\description{
This function queries the Income Limits API provided by
US Department of Housing and Urban Development.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_il("VA", year=c(2021))

hud_il("5100199999", year=c(2021))

hud_il("METRO47900M47900", year=c(2018))
}
}
\keyword{API}
\keyword{Income}
\keyword{Limits}
