% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_metropolitan}
\alias{hud_metropolitan}
\title{hud_metropolitan}
\usage{
hud_metropolitan()
}
\value{
A dataframe containing details of metropolitan areas in US
}
\description{
Get a list of all metropolitan areas in the US along with its name and CBSA code
}
\keyword{CBSA}
