% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_nation_states_territories}
\alias{hud_nation_states_territories}
\title{hud_nation_states_territories}
\usage{
hud_nation_states_territories(key = Sys.getenv("HUD_KEY"), to_tibble)
}
\arguments{
\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing details of all the states and territories
in the US.
}
\description{
Get a list of state and US territories
along with the corresponding fips code and
abbreviation.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_nation_states_territories()
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{States}
\keyword{Territories}
