% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_places}
\alias{hud_places}
\title{hud_places}
\usage{
hud_places(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all places.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing details of places in a state.
}
\description{
Get a list of all places in a state.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_places("CA")
hud_places("Virginia")
hud_places("51")
}
}
\keyword{places.}
