% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_find_most_recent_data.R
\name{hud_rec_cw_yr}
\alias{hud_rec_cw_yr}
\title{hud_rec_cw_yr}
\usage{
hud_rec_cw_yr(key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{key}{The API key for this user. You must go to HUD and sign up
for an account and request for an API key.}
}
\value{
The most recent year and quarter of available crosswalk files.
}
\description{
Ping the Crosswalk API provided by HUD User to determine the
most recently released files. This will only ping for the last two
years. This tests hud_cw_tract_zip(tract = 48201223100) as the
endpoint query.
}
\examples{
\dontrun{

hud_rec_cw_yr()

}
}
\seealso{
\itemize{
\item \code{\link[=hud_rec_cw_yr]{hud_rec_cw_yr()}}
\item \code{\link[=hud_rec_fmr_yr]{hud_rec_fmr_yr()}}
\item \code{\link[=hud_rec_il_yr]{hud_rec_il_yr()}}
}
}
