% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_find_most_recent_data.R
\name{hud_rec_fmr_yr}
\alias{hud_rec_fmr_yr}
\title{hud_rec_fmr_yr}
\usage{
hud_rec_fmr_yr()
}
\value{
The most recent year for the fmr files for state, county, and
metroarea queries.
}
\description{
Ping the Fair Markets Rent API provided by HUD User to
determine the most recently released files. This will only ping
for the last two years. Will return years for county and metroarea
resolution.
}
\examples{
\dontrun{

hud_rec_fmr_yr()

}
}
\seealso{
\itemize{
\item \code{\link[=hud_rec_cw_yr]{hud_rec_cw_yr()}}
\item \code{\link[=hud_rec_fmr_yr]{hud_rec_fmr_yr()}}
\item \code{\link[=hud_rec_il_yr]{hud_rec_il_yr()}}
}
}
