% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_key.R
\name{hud_set_key}
\alias{hud_set_key}
\title{hud_set_key}
\usage{
hud_set_key(key, in_wkdir = FALSE, in_home = FALSE)
}
\arguments{
\item{key}{key obtained at
https://www.huduser.gov/hudapi/public/register?comingfrom=1}

\item{in_wkdir}{set the key in the user's .RProfile in this directory.
Is defaulted to false.}

\item{in_home}{set the key in the user's HOME directory.
Is defaulted to false.}
}
\description{
A wrapper around Sys.getenv() to set HUD_KEY environment
variable. Allows the user to also set the key to their .RProfile
in working or HOME.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
hud_get_key()

hud_set_key("dwqkqodkqodqkqoqdq")
hud_get_key()
}
}
\seealso{
\itemize{
\item \code{\link[=hud_get_key]{hud_get_key()}}
\item \code{\link[=hud_set_key]{hud_set_key()}}
}
}
