% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_state_counties}
\alias{hud_state_counties}
\title{hud_state_counties}
\usage{
hud_state_counties(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all counties.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing all counties within a state
}
\description{
Get a list of all counties within a state.
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_counties("CA")
hud_state_counties("Virginia")
hud_state_counties("51")
}
}
\keyword{Counties}
