% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_misc.R
\name{hud_state_counties}
\alias{hud_state_counties}
\title{US Counties}
\usage{
hud_state_counties(
  state,
  key = Sys.getenv("HUD_KEY"),
  to_tibble = getOption("rhud_use_tibble", FALSE)
)
}
\arguments{
\item{state}{The state(s) to get all county(s).}

\item{key}{The key obtained from HUD
(US Department of Housing and Urban Development)
USER website.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing all counties within state(s).
}
\description{
Get a dataframe with details describing the county(s) located
within the queried state(s).
}
\examples{
\dontrun{
hud_state_counties("CA")
hud_state_counties("Virginia")
hud_state_counties("51")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{counties}
