% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_misc.R
\name{hud_state_metropolitan}
\alias{hud_state_metropolitan}
\title{hud_state_metropolitan}
\usage{
hud_state_metropolitan(state, key = Sys.getenv("HUD_KEY"), to_tibble)
}
\arguments{
\item{state}{The state to get all the metropolitan areas. Can be provided as
the full name, fip code or abbreviation.}

\item{key}{The API key for this user. You must go to HUD and sign up for an
account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing details of metropolitan areas in a state.
}
\description{
Get details for all metropolitan areas for queried states with
their name and CBSA code.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_metropolitan("VA")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{CBSA}
