% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_state_minor_civil_divisions}
\alias{hud_state_minor_civil_divisions}
\title{hud_state_minor_civil_divisions}
\usage{
hud_state_minor_civil_divisions(state, key = Sys.getenv("HUD_KEY"))
}
\arguments{
\item{state}{The state to get all MCD.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}
}
\value{
A dataframe containing details of minor civil divisions in a state.
}
\description{
Get a list of all minor civil divisions in a state
}
\examples{
\dontrun{
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_minor_civil_divisions("CA")
hud_state_minor_civil_divisions("Virginia")
hud_state_minor_civil_divisions("51")
}
}
\keyword{CBSA}
