% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_state_places}
\alias{hud_state_places}
\title{hud_state_places}
\usage{
hud_state_places(state, key = Sys.getenv("HUD_KEY"), to_tibble)
}
\arguments{
\item{state}{The state to get all places.}

\item{key}{The API key for this user. You must go to HUD and sign up for
an account and request for an API key.}

\item{to_tibble}{If TRUE, return the data in a tibble format
rather than a data frame.}
}
\value{
A dataframe containing details of places in a state.
}
\description{
Get a list of all places in a state.
}
\examples{
\dontrun{
library(rhud)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_state_places("CA")
hud_state_places("Virginia")
hud_state_places("51")
}
}
\seealso{
\itemize{
\item \code{\link[=hud_nation_states_territories]{hud_nation_states_territories()}}
\item \code{\link[=hud_state_metropolitan]{hud_state_metropolitan()}}
\item \code{\link[=hud_state_counties]{hud_state_counties()}}
\item \code{\link[=hud_state_places]{hud_state_places()}}
\item \code{\link[=hud_state_minor_civil_divisions]{hud_state_minor_civil_divisions()}}
}
}
\keyword{places.}
