% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudmisc.R
\name{hud_states}
\alias{hud_states}
\title{hud_states}
\usage{
hud_states()
}
\value{
A dataframe containing details of all the states in the US
}
\description{
Get a list of state along with the corresponding FIPs code and abbreviation
}
\keyword{States}
