% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huduser.R
\name{hudchas}
\alias{hudchas}
\title{hudchas}
\usage{
hudchas(
  type,
  stateId = "",
  entityId = "",
  year = "2014-2018",
  key = pkg.env$curr.key
)
}
\arguments{
\item{type}{Queries the data based off
1 - Nation
2 - State
3 - County
4 - MCD
5 - Place}

\item{stateId}{For types 2,3,4,5, you must provide a stateId. For 3,4,5}

\item{entityId}{For types 3,4,5, you must provide a fips code}

\item{year}{Gets the year that this data was recorded. Defaults to 2014-2018. There are specific year ranges
that are only accepted.
2014-2018
2013-2017
2012-2016
2011-2015
2010-2014
2009-2013
2008-2012
2007-2011
2006-2010}

\item{key}{The API key for this user. You must go to HUD and sign up for an account and request for an API key.}
}
\value{
This function returns a dataframe containing CHAS data for a particular state.
For more details about these measurements, go to https://www.huduser.gov/portal/dataset/fmr-api.html
}
\description{
This function queries the CHAS API provided by US Department of Housing and Urban Development
}
\keyword{(CHAS)}
\keyword{API}
\keyword{Affordability}
\keyword{Comprehensive}
\keyword{Housing}
\keyword{Strategy}
