% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{hudfmr}
\alias{hudfmr}
\title{hudfmr
This function queries the Free Markets Rent API provided by US Department of Housing and Urban Development}
\usage{
hudfmr(query, year = pkg.env$curr.year, key)
}
\arguments{
\item{query}{Can provide either FIPS code, CBSA code for metro areas, or state abbreviation.}

\item{year}{Gets the year that this data was recorded. Default is the latest year.}

\item{key}{The API key for this user. You must go to HUD and sign up for an account and request for an API key.}
}
\value{
This function returns a dataframe containing FREE MARKETS RENT data for a particular county or state.
For county level data, these measurements include the county_name, counties_msa, town_name, metro_status, metro_name,
smallarea_status, basicdata, Efficiency, One-Bedroom, Two-Bedroom, Three-Bedroom, Four-Bedroom, and year
For more details about these measurements, go to https://www.huduser.gov/portal/dataset/fmr-api.html
For state level data, these measurements will be the same as county level data, but will return a dataframe with
the individual measurements for each individual county within the state.
}
\description{
hudfmr
}
\examples{
fmr <- hudfmr(query = '22031', year = '2021', key = 'edf23jf834qd72nja')
fmr <- hudfmr(query = 'VA', year = '2021', key = 'edf23jf834qd72nja')
}
\keyword{API}
\keyword{Crosswalks}
