% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudexchange.R
\name{hudnsp}
\alias{hudnsp}
\title{hudnsp}
\usage{
hudnsp(file = 1)
}
\arguments{
\item{file}{The specific file needed.
1)NSP NSP Close-out_Admin Rept02b - with Act Start and End dates and ENV Status
2)
3)
4)
5)
6)
7)
8)}
}
\value{
A dataframe.
}
\description{
This will grab data from https://drgr.hud.gov/public/data_downloads.html?programName=NSP
and format it into a dataframe. It will unroll each field to remove NA values in cells. Currently this only supports getting access to the NSP NSP Grant Summary.
}
\details{
hudnsp
}
