% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudparallel.R
\name{parallelize_api_calls}
\alias{parallelize_api_calls}
\title{parallelize_api_calls}
\usage{
parallelize_api_calls(urls, cores = 1)
}
\arguments{
\item{urls}{A list of urls to query for.}

\item{cores}{The number of CPU cores to use when processing the requests.}
}
\value{
The response body provided by the API call.
}
\description{
Make API calls using multiple cores.
}
\keyword{parallel}
