% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rhud-package}
\alias{rhud-package}
\alias{rhud}
\title{rhud: A R Interface to the US Department of Housing and Urban
Development APIs}
\description{
rhud is a R library that allows users to easily download data from the
US Department of Housing and Urban Development APIs @ HUD User.
The data is defaulted to return measurements for the prior year
(the year 365 days ago from the current system date). For details about
this package as well as details about its python counterpart,
visit https://github.com/etam4260/rhud

The default return type of data is a dataframe.
To get tibbles instead, use: options(rhud_use_tibble = TRUE)
}
\author{
Emmet Tam
}
