% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_website_docs.R
\name{rhud_website}
\alias{rhud_website}
\title{rhud_website}
\usage{
rhud_website(website = c("github-pages", "github"))
}
\arguments{
\item{website}{The websites available.
\enumerate{
\item "github-pages"
\item "github"
}}
}
\description{
Quickly get documentation for the rhud package by opening up
the websites associated with it.
}
\examples{
\dontrun{
rhud_website("github-pages")
rhud_website("github")
rhud_website()
}
}
